commit dee730f
Author: Peter Kwekkeboom <peter.kwekkeboom@dekimo.com>
Date:   Tue Dec 12 07:35:51 2023 +0100

    change: add extra battery measurement to get the R and Z impedances.

commit b9074e9
Author: Peter Kwekkeboom <peter.kwekkeboom@dekimo.com>
Date:   Tue Dec 12 07:35:29 2023 +0100

    change: make helper function for battery measurement.

commit 83beda8
Author: Peter Kwekkeboom <peter.kwekkeboom@dekimo.com>
Date:   Tue Dec 12 07:34:32 2023 +0100

    change: cast value, not a real problem

+++ App/BatMon_BQ35100/bq35100.c
@@ -103,7 +103,7 @@ uint16_t bq35100_getScaledR(void)
 
   HAL_I2C_Mem_Read(bq35100Handle, BQ35100_ADDRESS, REG_SCALED_R, 1, data, 2, BQ35100_I2C_WAIT);
 
-  scaledR = (data[1]<<8 & 0xFF00) | (data[0] & 0x00FF);
+  scaledR = (int16_t)(data[1]<<8 & 0xFF00) | (data[0] & 0x00FF);
 
   return scaledR;
 }
@@ -120,7 +120,7 @@ uint16_t bq35100_getMeasuredZ(void)
 
   HAL_I2C_Mem_Read(bq35100Handle, BQ35100_ADDRESS, REG_MEASURED_Z, 1, data, 2, BQ35100_I2C_WAIT);
 
-  measuredZ = (data[1]<<8 & 0xFF00) | (data[0] & 0x00FF);
+  measuredZ = (int16_t)(data[1]<<8 & 0xFF00) | (data[0] & 0x00FF);
 
   return measuredZ;
 }
